Derive the modal value(s) for a set of data

This function takes a vector and returns one or mode values that represent the mode point of the data

Mode(x)

Arguments

x
a vector

Value

a vector containing one or more modal values for the input vector

Warning

Be careful the function does some basic error checking but the return to Mode(NA) is NA and a vector where the majority of entries are NA is also NA

Examples

Mode(sample(1:100, 1000, replace = TRUE))

#> [1] 73

Mode(mtcars$hp)

#> [1] 110 175 180

Mode(iris$Sepal.Length)

#> [1] 5