U.S. 2000 Election Data (short)

Data from a post-election survey following the year 2000 U.S. presidential elections. This is a subset from package `CHAID`.

USvoteS

Format

A data frame with 1000 observations on the following 6 variables.:

vote3

candidate voted for Gore or Bush

gender

gender, a factor with levels male and female

ager

age group, an ordered factor with levels 18-24 < 25-34 < 35-44 < 45-54 < 55-64 < 65+

empstat

status of employment, a factor with levels yes, no or retired

educr

status of education, an ordered factor with levels <HS < HS < >HS < College < Post Coll

marstat

status of living situation, a factor with levels married, widowed, divorced or never married

Source

https://r-forge.r-project.org/R/?group_id=343