Calculates eta squared for an AOV model using the Type II method

Takes an aov object and returns a standard AOV table with eta squared computed

neweta(MyAOV)

Arguments

MyAOV
a valid aov object such as those produced by aov(dv~iv1*iv2)

Value

Returns a tibble containing the AOV output similar to summary(aov(MyAOV)) but with eta squared computed and appended as an additional column

Details

There are three ways to compute eta squared this function only uses Type II

References

neweta function is a shortened and simplified verion of Dani Navarro's etaSquared

See also

Examples

neweta(aov(mpg~am*cyl, mtcars))

#> # A tibble: 4 x 8
#>   Source       Df `Sum Sq` `Mean Sq` `F value`      p sigstars `eta sq`
#>   <fct>     <int>    <dbl>     <dbl>     <dbl>  <dbl> <chr>       <dbl>
#> 1 am            1     37.0     37.0        4.3  0.048 *           0.033
#> 2 cyl           1    450.     450.        52    0     ***         0.399
#> 3 am:cyl        1     29.4     29.4        3.4  0.076 .           0.026
#> 4 Residuals    28    242.       8.64      NA   NA     <NA>        0.215